Broke the pipe file into smaller units
